move aws fetching to new file
