+++ components/studentPortal/layout/Sidebar.tsx
@@ -1,5 +1,4 @@
 import { useQuery } from "@apollo/client/react";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
@@ -9,6 +8,7 @@ import {
   FetchRoleData,
 } from "../../../graphql/studentPortal/users/fetchUserRole";
 import { useAuth } from "../../../lib/authContext";
+import { fetchProfilePicture } from "../../../pages/profile/profilePicturesClient";
 import {
   profileSelector,
   setCreatedAt,
@@ -24,11 +24,6 @@ import SkillifyCommandPalette from "./CommandPalette";
 import SidebarItem, { SidebarItemProps } from "./SidebarItem";
 
 // import aws s3
-import {
-  GetObjectCommand,
-  ListObjectsV2Command,
-  S3Client,
-} from "@aws-sdk/client-s3";
 
 export const Sidebar: React.FC = () => {
   const { goalApproaching } = useSelector(activePageSelector);
@@ -84,46 +79,15 @@ export const Sidebar: React.FC = () => {
     }
   }, [router.pathname]);
 
-  const fetchProfilePicture = async () => {
-    // fetch profile picture from aws s3 bucket
-    const s3 = new S3Client({
-      region: "us-east-1",
-      credentials: {
-        accessKeyId: process.env.NEXT_PUBLIC_AWS_ACCESS_KEY_ID,
-        secretAccessKey: process.env.NEXT_PUBLIC_AWS_SECRET_ACCESS_KEY,
-      },
-    });
-
-    const params = {
-      Bucket: "student-profile-pictures",
-      Prefix: `${user.uid}/`,
-    };
-    await s3
-      .send(new ListObjectsV2Command(params))
-      .then(async (res) => {
-        if (res.Contents && res.Contents.length > 0) {
-          const firstImage = res.Contents.filter((x) => !x.Key.endsWith("/"))[0]
-            .Key;
-          const imageUrl = await getSignedUrl(
-            s3,
-            new GetObjectCommand({
-              Bucket: "student-profile-pictures",
-              Key: firstImage,
-            })
-          );
+  useEffect(() => {
+    if (user) {
+      fetchProfilePicture(user.uid).then((imageUrl) => {
+        if (imageUrl !== null) {
           setUserProfileImage(imageUrl);
         } else {
           setUserProfileImage(user.photoURL);
         }
-      })
-      .catch((err) => {
-        console.log("Error fetching image from bucket", err);
       });
-  };
-
-  useEffect(() => {
-    if (user) {
-      fetchProfilePicture();
     }
   }, [user]);
